@@ -4,22 +4,24 @@
   "proxy": "http://localhost:5000",
   "private": true,
   "dependencies": {
-    "@mantine/core": "^3.6.7",
-    "@mantine/dates": "^3.6.7",
-    "@mantine/hooks": "^3.6.7",
-    "@mantine/modals": "^3.6.7",
-    "@mantine/notifications": "^3.6.7",
-    "@reduxjs/toolkit": "^1.7.2",
+    "@mantine/core": "^4.1.2",
+    "@mantine/dates": "^4.1.2",
+    "@mantine/form": "^4.1.2",
+    "@mantine/hooks": "^4.1.2",
+    "@mantine/modals": "^4.1.2",
+    "@mantine/notifications": "^4.1.2",
+    "@reduxjs/toolkit": "^1.8.1",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
-    "dayjs": "^1.10.7",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "^6.2.1",
-    "react-scripts": "5.0.0",
-    "react-timekeeper": "^2.2.1"
+    "dayjs": "^1.11.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-redux": "^7.2.8",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.0",
+    "react-timekeeper": "^2.2.1",
+    "recharts": "^2.1.9"
   },
   "scripts": {
     "start": "react-scripts start",
