@@ -6,7 +6,6 @@ import Router from './Router';
 
 function App() {
   const preferredColorScheme = useColorScheme();
-  const theme = useMantineTheme();
   return (
     <MantineProvider
       theme={{
@@ -14,15 +13,20 @@ function App() {
         colorScheme: preferredColorScheme,
         breakpoints: {
           xl: 1600
-        }
+        },
+        dateFormat: 'DD.MM.YYYY'
       }}
       styles={{
         Paper: (theme) => ({
           root: {
             backgroundColor: theme.colorScheme === 'dark' ? theme.colors.dark[6] : 'white',
+            padding: theme.spacing.xl,
+            borderRadius: theme.radius.md,
+            boxShadow: theme.shadows.sm
           }
         })
       }}
+      withCSSVariables
       withNormalizeCSS
       withGlobalStyles
     >
