@@ -3,8 +3,7 @@ import { Navbar as MantineNavbar, Box } from '@mantine/core'
 import NavLink from './NavLink'
 import { IoStorefrontOutline, IoHomeOutline, IoAlbumsOutline, IoBriefcaseOutline } from 'react-icons/io5'
 import NavUser from './NavUser'
-function Navbar(props) {
-    const {withBurger, menuOpened, toggleMenu, ...navbarProps} = props;
+function Navbar({ withBurger, menuOpened, toggleMenu, ...navbarProps }) {
     return (
         <MantineNavbar {...navbarProps}>
             <MantineNavbar.Section grow>
@@ -15,7 +14,7 @@ function Navbar(props) {
             </MantineNavbar.Section>
             <MantineNavbar.Section>
                 <NavUser />
-                <Box style={{ fontSize: '12px', margin: '10px 0', paddingLeft: '12px'}}>Studentský zápočtový projekt - Jan Joska</Box>
+                <Box style={{ fontSize: '12px', margin: '10px 0', paddingLeft: '12px' }}>Studentský zápočtový projekt - Jan Joska</Box>
             </MantineNavbar.Section>
         </MantineNavbar>
     )
